requirements.txt
flask
tensorflow
pillow
numpy
scipy
pyopenssl
gunicorn
google-generativeai